REF: move PandasDtype to dtypes.dtypes
