+++ pandas/core/arrays/__init__.py
@@ -10,7 +10,7 @@ from pandas.core.arrays.floating import FloatingArray
 from pandas.core.arrays.integer import IntegerArray
 from pandas.core.arrays.interval import IntervalArray
 from pandas.core.arrays.masked import BaseMaskedArray
-from pandas.core.arrays.numpy_ import PandasArray, PandasDtype
+from pandas.core.arrays.numpy_ import PandasArray
 from pandas.core.arrays.period import PeriodArray, period_array
 from pandas.core.arrays.sparse import SparseArray
 from pandas.core.arrays.string_ import StringArray
@@ -28,7 +28,6 @@ __all__ = [
     "IntegerArray",
     "IntervalArray",
     "PandasArray",
-    "PandasDtype",
     "PeriodArray",
     "period_array",
     "SparseArray",
