+++ pandas/core/arrays/numpy_.py
@@ -1,7 +1,7 @@
 from __future__ import annotations
 
 import numbers
-from typing import Optional, Tuple, Type, Union
+from typing import Optional, Tuple, Union
 
 import numpy as np
 from numpy.lib.mixins import NDArrayOperatorsMixin
@@ -10,108 +10,13 @@ from pandas._libs import lib
 from pandas._typing import Dtype, NpDtype, Scalar
 from pandas.compat.numpy import function as nv
 
-from pandas.core.dtypes.dtypes import ExtensionDtype
+from pandas.core.dtypes.dtypes import PandasDtype
 from pandas.core.dtypes.missing import isna
 
 from pandas.core import nanops, ops
 from pandas.core.arraylike import OpsMixin
 from pandas.core.arrays._mixins import NDArrayBackedExtensionArray
 from pandas.core.strings.object_array import ObjectStringArrayMixin
-
-
-class PandasDtype(ExtensionDtype):
-    """
-    A Pandas ExtensionDtype for NumPy dtypes.
-
-    .. versionadded:: 0.24.0
-
-    This is mostly for internal compatibility, and is not especially
-    useful on its own.
-
-    Parameters
-    ----------
-    dtype : object
-        Object to be converted to a NumPy data type object.
-
-    See Also
-    --------
-    numpy.dtype
-    """
-
-    _metadata = ("_dtype",)
-
-    def __init__(self, dtype: Optional[NpDtype]):
-        self._dtype = np.dtype(dtype)
-
-    def __repr__(self) -> str:
-        return f"PandasDtype({repr(self.name)})"
-
-    @property
-    def numpy_dtype(self) -> np.dtype:
-        """
-        The NumPy dtype this PandasDtype wraps.
-        """
-        return self._dtype
-
-    @property
-    def name(self) -> str:
-        """
-        A bit-width name for this data-type.
-        """
-        return self._dtype.name
-
-    @property
-    def type(self) -> Type[np.generic]:
-        """
-        The type object used to instantiate a scalar of this NumPy data-type.
-        """
-        return self._dtype.type
-
-    @property
-    def _is_numeric(self) -> bool:
-        # exclude object, str, unicode, void.
-        return self.kind in set("biufc")
-
-    @property
-    def _is_boolean(self) -> bool:
-        return self.kind == "b"
-
-    @classmethod
-    def construct_from_string(cls, string: str) -> PandasDtype:
-        try:
-            dtype = np.dtype(string)
-        except TypeError as err:
-            if not isinstance(string, str):
-                msg = f"'construct_from_string' expects a string, got {type(string)}"
-            else:
-                msg = f"Cannot construct a 'PandasDtype' from '{string}'"
-            raise TypeError(msg) from err
-        return cls(dtype)
-
-    @classmethod
-    def construct_array_type(cls) -> Type[PandasArray]:
-        """
-        Return the array type associated with this dtype.
-
-        Returns
-        -------
-        type
-        """
-        return PandasArray
-
-    @property
-    def kind(self) -> str:
-        """
-        A character code (one of 'biufcmMOSUV') identifying the general kind of data.
-        """
-        return self._dtype.kind
-
-    @property
-    def itemsize(self) -> int:
-        """
-        The element size of this data-type object.
-        """
-        return self._dtype.itemsize
 
 
 class PandasArray(
