+++ pandas/core/dtypes/dtypes.py
@@ -31,7 +31,7 @@ from pandas._libs.tslibs import (
     to_offset,
     tz_compare,
 )
-from pandas._typing import Dtype, DtypeObj, Ordered
+from pandas._typing import Dtype, DtypeObj, NpDtype, Ordered
 
 from pandas.core.dtypes.base import ExtensionDtype, register_extension_dtype
 from pandas.core.dtypes.generic import ABCCategoricalIndex, ABCIndex
@@ -41,7 +41,12 @@ if TYPE_CHECKING:
     import pyarrow
 
     from pandas import Categorical
-    from pandas.core.arrays import DatetimeArray, IntervalArray, PeriodArray
+    from pandas.core.arrays import (
+        DatetimeArray,
+        IntervalArray,
+        PandasArray,
+        PeriodArray,
+    )
 
 str_type = str
 
@@ -1230,3 +1235,103 @@ class IntervalDtype(PandasExtensionDtype):
         if common == object:
             return np.dtype(object)
         return IntervalDtype(common, closed=closed)
+
+
+class PandasDtype(ExtensionDtype):
+    """
+    A Pandas ExtensionDtype for NumPy dtypes.
+
+    .. versionadded:: 0.24.0
+
+    This is mostly for internal compatibility, and is not especially
+    useful on its own.
+
+    Parameters
+    ----------
+    dtype : object
+        Object to be converted to a NumPy data type object.
+
+    See Also
+    --------
+    numpy.dtype
+    """
+
+    _metadata = ("_dtype",)
+
+    def __init__(self, dtype: Optional[Union[NpDtype, PandasDtype]]):
+        if isinstance(dtype, PandasDtype):
+            # make constructor univalent
+            dtype = dtype.numpy_dtype
+        self._dtype = np.dtype(dtype)
+
+    def __repr__(self) -> str:
+        return f"PandasDtype({repr(self.name)})"
+
+    @property
+    def numpy_dtype(self) -> np.dtype:
+        """
+        The NumPy dtype this PandasDtype wraps.
+        """
+        return self._dtype
+
+    @property
+    def name(self) -> str:
+        """
+        A bit-width name for this data-type.
+        """
+        return self._dtype.name
+
+    @property
+    def type(self) -> Type[np.generic]:
+        """
+        The type object used to instantiate a scalar of this NumPy data-type.
+        """
+        return self._dtype.type
+
+    @property
+    def _is_numeric(self) -> bool:
+        # exclude object, str, unicode, void.
+        return self.kind in set("biufc")
+
+    @property
+    def _is_boolean(self) -> bool:
+        return self.kind == "b"
+
+    @classmethod
+    def construct_from_string(cls, string: str) -> PandasDtype:
+        try:
+            dtype = np.dtype(string)
+        except TypeError as err:
+            if not isinstance(string, str):
+                msg = f"'construct_from_string' expects a string, got {type(string)}"
+            else:
+                msg = f"Cannot construct a 'PandasDtype' from '{string}'"
+            raise TypeError(msg) from err
+        return cls(dtype)
+
+    @classmethod
+    def construct_array_type(cls) -> Type[PandasArray]:
+        """
+        Return the array type associated with this dtype.
+
+        Returns
+        -------
+        type
+        """
+        from pandas.core.arrays import PandasArray
+
+        return PandasArray
+
+    @property
+    def kind(self) -> str:
+        """
+        A character code (one of 'biufcmMOSUV') identifying the general kind of data.
+        """
+        return self._dtype.kind
+
+    @property
+    def itemsize(self) -> int:
+        """
+        The element size of this data-type object.
+        """
+        return self._dtype.itemsize
