+++ pandas/tests/extension/test_numpy.py
@@ -16,11 +16,11 @@ be added to the array-specific tests in `pandas/tests/arrays/`.
 import numpy as np
 import pytest
 
-from pandas.core.dtypes.dtypes import ExtensionDtype
+from pandas.core.dtypes.dtypes import ExtensionDtype, PandasDtype
 
 import pandas as pd
 import pandas._testing as tm
-from pandas.core.arrays.numpy_ import PandasArray, PandasDtype
+from pandas.core.arrays.numpy_ import PandasArray
 
 from . import base
